week_12/weekend_homework_music_shop/music_shop_solution/src/main/java/music_shop/MusicStand.java
package music_shop;

import behaviours.ISell;

public class MusicStand implements ISell {

    private String colour;
    private int buyPrice;
    private int sellPrice;

    public MusicStand(String colour, int buyPrice, int sellPrice) {
        this.colour = colour;
        this.buyPrice = buyPrice;
        this.sellPrice = sellPrice;
    }

    public String getColour() {
        return this.colour;
    }

    public int getBuyPrice() {
        return this.buyPrice;
    }

    public int getSellPrice() {
        return this.sellPrice;
    }

    public int calculateMarkup() {
        return this.sellPrice - this.buyPrice;
    }
}
